assettransfer with transfer address in tx
assettransfer should show sender and transfer recipient in the json payload, otherwise its really hard to know who the sender/recipient was for an asset transfer as there is really no inference on where it came from or where its going unless specifically shown in the tx. The other thing is we want to know at the time of the tx who it sent and where it sent from for historically accurate results.