refactor getsyscoinburndata, not needed for sysburntxtojson
sysburntxtojson can call getsyscoindata because there should only be one arg in the opreturn data field which would be the eth address.

there was a bug in the txtojson calls for asset that wouldn't return the changed data because it depended on the current database asset which ofcourse after a confirmation would match the transaction payload and not display the changeset, instead we show the changeset based on the payload field being empty or not. We also enforce precision and witnessaddress fields to ensure that they are not tampered with on transaction inclusion showing bad json entries.