remove Dosy checks around txroot validation
we want to avoid blocks from being flagged as invalid and thus avoid reprocessing them as txroot checks are sometimes intermittent (they are dependent on the relayer). So we flag as error instead of invalid which at the tx level it rejects and doesn't ban and on the block level it rejects but reprocesses again (assuming relayer feeds the correct information at some time, will reprocess the block if the transaction was legit).

We should look at this later and figure out when Dosy checks are needed like thresholds where txroots can fall into after looking at some statistics of borderline rejections. Once we have those stats we can then add Dosy checks around transactions that are obviously wrong versus borderline.