fix checkblock bug, remove arrival-times check
the arrival-times check won't work and is really a far fetched scenario likely invalid. The checkblock wasn't actually checking for the syscoin tx root there properly because of fChecked being set in acceptblock() so we check  if pow and merkle root are both checked. We also want to return state.error() on compact block checks before relay and state.invalid() if the block is trying to connect so we don't get into situation as described in #368

We need this check here and in checksyscoinmint which does enforcing through banning and invalid because checksyscoinmint is also called from accepttomemorypool() when new txs are validated as recieved from mempool.